+++ src/kafunk/Producer.fs
@@ -502,13 +502,14 @@ module Producer =
       do! state.partitionQueues.[batch.partition] batch
       return! batch.rep |> IVar.getWithTimeout p.batchTimeout (fun _ -> Failure (ProducerError.BatchTimeoutError batch)) }
 
-  let private splitBySize (maxBatchSize:int) (batch:ProducerMessage seq) =
+  let private splitBySize (maxBatchSize:int) (batch:ProducerMessage ResizeArray) =
     // NB: if the size of an individual message is larger than the max, then this
     // would still add it to the batch and potentially cause the server to return an error
     let batches = ResizeArray<_>()
     let currentBatch = ResizeArray<_>()
     let mutable currentBatchSize = 0
-    for pm in batch do
+    for i = 0 to batch.Count - 1 do
+      let pm = batch.[i]
       let messageSize = ProducerMessage.size pm
       if (currentBatchSize + messageSize) >= maxBatchSize then 
         if currentBatch.Count > 0 then